# zhiyin-src/template/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_ModuleResult_": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModuleResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "description": "链路追踪 id，由 BFF 生成并回写 X-Trace-Id 响应头；日志排查用",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[ModuleResult]",
        "type": "object"
      },
...
      "ApiResponse_list_ModuleView__": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ModuleView"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "description": "链路追踪 id，由 BFF 生成并回写 X-Trace-Id 响应头；日志排查用",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[list[ModuleView]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModuleAction": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "action"
        ],
        "title": "ModuleAction",
        "type": "object"
      },
...
      "ModuleFlowPublish": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "ModuleFlowPublish",
        "type": "object"
      },
...
      "ModuleInvoke": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "input": {
            "additionalProperties": true,
            "title": "Input",
            "type": "object"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "title": "ModuleInvoke",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "职引 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/app/modules": {
      "get": {
        "operationId": "user_modules_api_v1_app_modules_get",
        "parameters": [
          {
            "in": "query",
            "name": "surface",
            "required": false,
            "schema": {
              "default": "application",
              "enum": [
                "application",
                "conversation"
              ],
              "title": "Surface",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ModuleView__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "User Modules",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/app/modules/{module_id}/actions": {
      "post": {
        "operationId": "module_action_api_v1_app_modules__module_id__actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Module Action",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/app/modules/{module_id}/data": {
      "get": {
        "operationId": "module_data_api_v1_app_modules__module_id__data_get",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Module Data",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/app/modules/{module_id}/invoke": {
      "post": {
        "operationId": "invoke_module_api_v1_app_modules__module_id__invoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleInvoke"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invoke Module",
        "tags": [
          "modules"
        ]
      }
    },
...
    "/api/v1/app/workflows/{workflow_id}/run": {
      "post": {
        "operationId": "run_workflow_api_v1_app_workflows__workflow_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleFlowRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleFlowRun_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Workflow",
        "tags": [
          "modules"
        ]
      }
    },
...
    "/api/v1/developer/executor/{job_id}": {
      "post": {
        "operationId": "update_release_api_v1_developer_executor__job_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutorUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReleaseJob_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Release",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/developer/executor/{job_id}/authorization": {
      "get": {
        "operationId": "release_authorization_api_v1_developer_executor__job_id__authorization_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Authorization",
        "tags": [
          "developer-projects"
        ]
      }
    },
...
    "/api/v1/developer/modules/{module_id}/policy": {
      "put": {
        "operationId": "module_policy_api_v1_developer_modules__module_id__policy_put",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModulePolicy"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModulePolicy_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Module Policy",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/developer/modules/{module_id}/preview": {
      "post": {
        "operationId": "preview_module_api_v1_developer_modules__module_id__preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModulePreview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Module",
        "tags": [
          "modules"
        ]
      }
    },
...
    "/api/v1/developer/releases": {
      "get": {
        "operationId": "release_jobs_api_v1_developer_releases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ReleaseJob__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Release Jobs",
        "tags": [
          "modules"
        ]
      },
      "post": {
        "operationId": "request_release_api_v1_developer_releases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReleaseJob_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Release",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/developer/releases/{job_id}/approve": {
      "post": {
        "operationId": "approve_release_api_v1_developer_releases__job_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReleaseJob_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Release",
        "tags": [
          "modules"
        ]
      }
    },
...
    "/api/v1/developer/workflows": {
      "get": {
        "operationId": "module_workflows_api_v1_developer_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ModuleFlowView__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Module Workflows",
        "tags": [
          "modules"
        ]
      },
      "post": {
        "operationId": "save_workflow_api_v1_developer_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleFlowDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleFlowView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Workflow",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/developer/workflows/{workflow_id}/publish": {
      "post": {
        "operationId": "publish_workflow_api_v1_developer_workflows__workflow_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleFlowPublish"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleFlowView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Workflow",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/developer/workflows/{workflow_id}/run": {
      "post": {
        "operationId": "preview_workflow_api_v1_developer_workflows__workflow_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleFlowRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleFlowRun_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Workflow",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/developer/workflows/{workflow_id}/runs": {
      "get": {
        "operationId": "workflow_runs_api_v1_developer_workflows__workflow_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ModuleFlowRun__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workflow Runs",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/v1/developer/workflows/{workflow_id}/unpublish": {
      "post": {
        "operationId": "unpublish_workflow_api_v1_developer_workflows__workflow_id__unpublish_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleFlowPublish"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModuleFlowView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unpublish Workflow",
        "tags": [
          "modules"
        ]
      }
    },
...
  }
}